pyproject.toml
[tool.poetry]
name = "fastzero"
version = "0.1.0"
description = "Curso FastApi dunossauro"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "0.111.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
pydantic-settings = "^2.0.3"
pydantic = {extras = ["email"], version = "^2.4.2"}
alembic = "^1.12.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
ruff = "^0.4.2"
httpx = "^0.25.0"
isort = "^5.12.0"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
preview = true
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'uvicorn fastzero.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastzero -vv'
post_test = 'coverage html'